@@ -28,12 +28,12 @@ def siteUrl = 'https://github.com/miteshpithadiya/SearchableSpinner'
 def gitUrl = 'https://github.com/miteshpithadiya/SearchableSpinner.git'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.2"
+    compileSdkVersion 30
+    buildToolsVersion '30.0.2'
 
     defaultConfig {
         minSdkVersion 15
-        targetSdkVersion 23
+        targetSdkVersion 30
         versionCode 8
         versionName "1.3.1"
     }
@@ -63,40 +63,7 @@ bintray {
     }
 }
 
-apply plugin: 'com.github.dcendents.android-maven'
-
 group = "com.toptoche.searchablespinner"
-
-install {
-    repositories.mavenInstaller {
-        pom {
-            project {
-                packaging 'aar'
-                name 'A simple spinner with an extra ability to search for the items.'
-                url siteUrl
-                licenses {
-                    license {
-                        name 'The Apache Software License, Version 2.0'
-                        url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-                    }
-                }
-                developers {
-                    developer {
-                        id 'miteshpithadiya'
-                        name 'Mitesh Pithadiya'
-                        email 'miteshvanp@hotmail.com'
-                    }
-                }
-                scm {
-                    connection gitUrl
-                    developerConnection gitUrl
-                    url siteUrl
-
-                }
-            }
-        }
-    }
-}
 
 dependencies {
 }
@@ -118,8 +85,4 @@ task javadocJar(type: Jar, dependsOn: javadoc) {
 artifacts {
     archives javadocJar
     archives sourcesJar
-}
-
-task findConventions << {
-    println project.getConvention()
 }
